+++ packages/logic-core/src/data/events/onSetGrid.ts
@@ -15,3 +15,23 @@ export function handlesSetGrid<T extends Instruction>(
 ): val is T & SetGridHandler {
   return isEventHandler(val, 'onSetGrid');
 }
+
+export function invokeSetGrid(
+  oldGrid: GridData,
+  newGrid: GridData,
+  solution: GridData | null
+) {
+  newGrid.symbols.forEach(list => {
+    list.forEach(symbol => {
+      if (handlesSetGrid(symbol)) {
+        newGrid = symbol.onSetGrid(oldGrid, newGrid, solution);
+      }
+    });
+  });
+  newGrid.rules.forEach(rule => {
+    if (handlesSetGrid(rule)) {
+      newGrid = rule.onSetGrid(oldGrid, newGrid, solution);
+    }
+  });
+  return newGrid;
+}