+++ packages/logic-core/src/data/rules/perfectionRule.ts
@@ -1,10 +1,18 @@
 import { AnyConfig } from '../config.js';
 import GridData from '../grid.js';
-import { Color, GridState, RuleState, State, Position } from '../primitives.js';
+import {
+  Color,
+  GridState,
+  RuleState,
+  State,
+  Position,
+  Mode,
+} from '../primitives.js';
 import Rule, { SearchVariant } from './rule.js';
 import CustomIconSymbol from '../symbols/customIconSymbol.js';
 import { SetGridHandler } from '../events/onSetGrid.js';
 import { FinalValidationHandler } from '../events/onFinalValidation.js';
+import validateGrid from '../validate.js';
 
 export default class PerfectionRule
   extends Rule
@@ -16,13 +24,18 @@ export default class PerfectionRule
     )
   );
 
-  private static readonly SEARCH_VARIANTS = []; // this rule is not searchable
+  private static readonly SEARCH_VARIANTS = [
+    new PerfectionRule().searchVariant(),
+  ];
 
   /**
    * **Quest for Perfection: cell colors are final**
+   *
+   * @param editor - whether to enable editor mode. This field is automatically set by the editor.
    */
-  public constructor() {
+  public constructor(public readonly editor = false) {
     super();
+    this.editor = editor;
   }
 
   public get id(): string {
@@ -51,6 +64,17 @@ export default class PerfectionRule
 
   public get isSingleton(): boolean {
     return true;
+  }
+
+  public modeVariant(mode: Mode): Rule | null {
+    // only allow this rule in perfection mode
+    if (this.editor === (mode === Mode.Create)) {
+      return this;
+    } else if (mode === Mode.Create) {
+      return this.copyWith({ editor: true });
+    } else {
+      return this.copyWith({ editor: false });
+    }
   }
 
   public validateGrid(grid: GridData): RuleState {
@@ -117,6 +141,10 @@ export default class PerfectionRule
     return grid;
   }
 
+  private isValid(grid: GridData, solution: GridData | null): boolean {
+    return validateGrid(grid, solution).final !== State.Error;
+  }
+
   /**
    * Force all tiles to be fixed.
    *
@@ -127,20 +155,16 @@ export default class PerfectionRule
     newGrid: GridData,
     solution: GridData | null
   ): GridData {
-    if (!solution) {
-      return this.fixTiles(newGrid);
-    }
-    if (
-      !oldGrid.solutionMatches(solution) &&
-      !newGrid.solutionMatches(solution)
-    ) {
+    if (this.editor) return newGrid;
+
+    if (!this.isValid(oldGrid, solution) && !this.isValid(newGrid, solution)) {
       return this.fixTiles(oldGrid);
     }
     return this.fixTiles(newGrid);
   }
 
-  public copyWith(_: object): this {
-    return new PerfectionRule() as this;
+  public copyWith({ editor }: { editor?: boolean }): this {
+    return new PerfectionRule(editor ?? this.editor) as this;
   }
 }