+++ src/client/configs/parts/NullableGridConfig.tsx
@@ -1,9 +1,9 @@
-import { memo, useState } from 'react';
+import { memo, useRef } from 'react';
 import { ConfigType, NullableGridConfig } from '@logic-pad/core/data/config';
 import Configurable from '@logic-pad/core/data/configurable';
 import GridData from '@logic-pad/core/data/grid';
 import { FiExternalLink } from 'react-icons/fi';
-import GridEditorModal from './GridEditorModal';
+import GridEditorModal, { GridEditorRef } from './GridEditorModal';
 import { FaTrashCan } from 'react-icons/fa6';
 
 export interface NullableGridConfigProps {
@@ -20,7 +20,7 @@ export default memo(function GridConfig({
   const grid = configurable[
     config.field as keyof typeof configurable
   ] as unknown as GridData | null;
-  const [open, setOpen] = useState(false);
+  const editorRef = useRef<GridEditorRef>(null);
 
   return (
     <div className="flex p-2 justify-between items-center">
@@ -55,17 +55,15 @@ export default memo(function GridConfig({
           <button
             type="button"
             className="btn justify-start flex-nowrap flex"
-            onClick={() => setOpen(true)}
+            onClick={() => editorRef.current?.open(grid)}
           >
             Open editor
             <FiExternalLink size={24} />
           </button>
 
           <GridEditorModal
-            grid={grid}
-            setGrid={grid => setConfig?.(config.field, grid)}
-            open={open}
-            setOpen={setOpen}
+            ref={editorRef}
+            onChange={grid => setConfig?.(config.field, grid)}
           />
         </div>
       )}