+++ src/client/configs/parts/SolvePathEditorModal.tsx
@@ -1,4 +1,10 @@
-import { memo, useEffect, useState } from 'react';
+import {
+  forwardRef,
+  memo,
+  useEffect,
+  useImperativeHandle,
+  useState,
+} from 'react';
 import { cn } from '../../uiHelper';
 import EmbedContext from '../../contexts/EmbedContext';
 import GridContext, { GridConsumer } from '../../contexts/GridContext';
@@ -7,69 +13,37 @@ import GridStateContext from '../../contexts/GridStateContext';
 import { Color, Position } from '@logic-pad/core/data/primitives';
 import PerfectionRule from '@logic-pad/core/data/rules/perfectionRule';
 import PerfectionScreen from '../../screens/PerfectionScreen';
-import ForesightRule, {
-  instance as foresightInstance,
-} from '@logic-pad/core/data/rules/foresightRule';
-import {
-  SolvePathConsumer,
-  useSolvePath,
-} from '../../contexts/SolvePathContext';
+import { instance as foresightInstance } from '@logic-pad/core/data/rules/foresightRule';
 import EditContext from '../../contexts/EditContext';
 import { useDelta } from 'react-delta-hooks';
 import { GridData } from '@logic-pad/core/index';
+import { Puzzle, PuzzleMetadata } from '@logic-pad/core/data/puzzle';
+import { invokeSetGrid } from '@logic-pad/core/data/events/onSetGrid';
 
 export interface SolvePathEditorModalProps {
-  solvePath: Position[];
-  setSolvePath: (solvePath: Position[]) => void;
-  open: boolean;
-  setOpen: (open: boolean) => void;
+  onChange: (solvePath: Position[]) => void;
 }
 
-function EmbedLoader({ solvePath }: { solvePath: Position[] }) {
-  const { setSolvePath } = useSolvePath();
-  useEffect(() => {
-    setSolvePath(solvePath);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-  return null;
-}
-
-function SolvePathUpdater({
-  oldSolvePath,
-  setSolvePath,
-}: {
-  oldSolvePath: Position[];
-  setSolvePath: (solvePath: Position[]) => void;
-}) {
-  const { solvePath: newSolvePath } = useSolvePath();
-  useEffect(() => {
-    if (newSolvePath !== oldSolvePath) setSolvePath(newSolvePath);
-  }, [newSolvePath, oldSolvePath, setSolvePath]);
-  return null;
+export interface SolvePathEditorRef {
+  open: (value: Position[], grid: GridData, metadata: PuzzleMetadata) => void;
 }
 
 function prepareGrid(
   grid: GridData,
-  respectSolvePath: boolean
+  solvePath: Position[] | null
 ): { grid: GridData; solution: GridData | null } {
   const newGrid = grid.withRules(rules => [
     new PerfectionRule(),
     ...rules.filter(r => r.id !== foresightInstance.id),
   ]);
-  if (respectSolvePath) {
-    const foresight = grid.findRule(r => r.id === foresightInstance.id) as
-      | ForesightRule
-      | undefined;
-    if (!foresight) {
-      return prepareGrid(newGrid, false);
-    }
+  if (solvePath) {
     const resetGrid = newGrid.withTiles(tiles =>
       tiles.map((row, y) =>
         row.map((tile, x) => {
           if (
             !tile.exists ||
             tile.fixed ||
-            foresight.solvePath.some(p => p.x === x && p.y === y)
+            solvePath.some(p => p.x === x && p.y === y)
           ) {
             return tile;
           } else {
@@ -79,127 +53,133 @@ function prepareGrid(
       )
     );
     if (resetGrid.colorEquals(newGrid)) {
-      return { grid: newGrid, solution: null };
+      return { grid: invokeSetGrid(resetGrid, newGrid, null), solution: null };
     } else {
-      return { grid: resetGrid, solution: newGrid };
+      return {
+        grid: invokeSetGrid(resetGrid, resetGrid, null),
+        solution: newGrid,
+      };
     }
   } else {
     const resetGrid = newGrid.resetTiles();
     if (resetGrid.colorEquals(newGrid)) {
-      return { grid: newGrid, solution: null };
+      return { grid: invokeSetGrid(resetGrid, newGrid, null), solution: null };
     } else {
-      return { grid: resetGrid, solution: newGrid };
+      return {
+        grid: invokeSetGrid(resetGrid, resetGrid, null),
+        solution: newGrid,
+      };
     }
   }
 }
 
-function lockGrid(grid: GridData): GridData {
-  return grid.withTiles(tiles =>
-    tiles.map(row =>
-      row.map(tile =>
-        tile.exists && tile.color !== Color.Gray ? tile.withFixed(true) : tile
-      )
-    )
-  );
-}
+export default memo(
+  forwardRef<SolvePathEditorRef, SolvePathEditorModalProps>(
+    function SolvePathEditorModal(
+      { onChange }: SolvePathEditorModalProps,
+      ref
+    ) {
+      /**
+       * initialState also specifies the open state of the modal.
+       */
+      const [initialState, setInitialState] = useState<Puzzle | null>(null);
+      const [tempSolvePath, setTempSolvePath] = useState<Position[]>([]);
 
-export default memo(function SolvePathEditorModal({
-  solvePath,
-  setSolvePath,
-  open,
-  setOpen,
-}: SolvePathEditorModalProps) {
-  const [tempSolvePath, setTempSolvePath] = useState(solvePath);
+      useImperativeHandle(ref, () => ({
+        open: (value: Position[], grid: GridData, metadata: PuzzleMetadata) => {
+          const { grid: newGrid, solution } = prepareGrid(grid, value);
+          setInitialState({ ...metadata, grid: newGrid, solution });
+          setTempSolvePath(value);
+        },
+      }));
 
-  const openDelta = useDelta(open);
-  useEffect(() => {
-    if (!openDelta) return;
-    if (!openDelta.curr && openDelta.prev) {
-      setSolvePath(tempSolvePath);
-    }
-  }, [openDelta, setSolvePath, tempSolvePath]);
+      const openDelta = useDelta(initialState);
+      useEffect(() => {
+        if (!openDelta) return;
+        if (!openDelta.curr && openDelta.prev) {
+          onChange(tempSolvePath);
+        }
+      }, [onChange, openDelta, tempSolvePath]);
 
-  return (
-    <dialog id="gridModal" className={cn('modal', open && 'modal-open')}>
-      <div className="modal-box w-[calc(100%-4rem)] h-full max-w-none bg-neutral text-neutral-content">
-        <form method="dialog">
-          <button
-            aria-label="Close dialog"
-            className="btn btn-sm btn-circle btn-ghost absolute right-2 top-2"
-            onClick={() => setOpen(false)}
-          >
-            ✕
-          </button>
-        </form>
-        {open && (
-          <GridConsumer>
-            {({ grid: outerGrid, metadata }) => {
-              return (
-                <EmbedContext>
-                  <DisplayContext>
-                    <EditContext>
-                      <GridStateContext>
-                        <GridContext
-                          grid={() =>
-                            lockGrid(prepareGrid(outerGrid, true).grid)
-                          }
-                          solution={() => prepareGrid(outerGrid, true).solution}
-                          metadata={metadata}
+      return (
+        <dialog
+          id="gridModal"
+          className={cn('modal', initialState && 'modal-open')}
+        >
+          <div className="modal-box w-[calc(100%-4rem)] h-full max-w-none bg-neutral text-neutral-content">
+            <form method="dialog">
+              <button
+                aria-label="Close dialog"
+                className="btn btn-sm btn-circle btn-ghost absolute right-2 top-2"
+                onClick={() => setInitialState(null)}
+              >
+                ✕
+              </button>
+            </form>
+            {initialState && (
+              <EmbedContext name="solve-path-modal">
+                <DisplayContext>
+                  <EditContext>
+                    <GridStateContext>
+                      <GridContext
+                        initialGrid={initialState.grid}
+                        initialSolution={initialState.solution}
+                        initialMetadata={() => {
+                          const {
+                            grid: _1,
+                            solution: _2,
+                            ...metadata
+                          } = initialState;
+                          return metadata;
+                        }}
+                      >
+                        <PerfectionScreen
+                          solvePath={tempSolvePath}
+                          setSolvePath={setTempSolvePath}
                         >
-                          <PerfectionScreen>
-                            <EmbedLoader solvePath={solvePath} />
-                            <SolvePathUpdater
-                              oldSolvePath={tempSolvePath}
-                              setSolvePath={setTempSolvePath}
-                            />
-                            <SolvePathConsumer>
-                              {({ setSolvePath }) => (
-                                <GridConsumer>
-                                  {({ setGrid: setInnerGrid }) => {
-                                    return (
-                                      <>
-                                        <button
-                                          type="button"
-                                          className="btn"
-                                          onClick={() => {
-                                            setInnerGrid(
-                                              prepareGrid(outerGrid, false).grid
-                                            );
-                                            setTempSolvePath([]);
-                                            setSolvePath([]);
-                                          }}
-                                        >
-                                          Reset progress
-                                        </button>
-                                        <button
-                                          type="button"
-                                          className="btn btn-primary"
-                                          onClick={() => {
-                                            setOpen(false);
-                                          }}
-                                        >
-                                          Save and exit
-                                        </button>
-                                      </>
-                                    );
-                                  }}
-                                </GridConsumer>
-                              )}
-                            </SolvePathConsumer>
-                          </PerfectionScreen>
-                        </GridContext>
-                      </GridStateContext>
-                    </EditContext>
-                  </DisplayContext>
-                </EmbedContext>
-              );
-            }}
-          </GridConsumer>
-        )}
-      </div>
-      <form method="dialog" className="modal-backdrop bg-neutral/55"></form>
-    </dialog>
-  );
-});
+                          <GridConsumer>
+                            {({ setGrid: setInnerGrid }) => {
+                              return (
+                                <>
+                                  <button
+                                    type="button"
+                                    className="btn"
+                                    onClick={() => {
+                                      const { grid, solution } = prepareGrid(
+                                        initialState.grid,
+                                        []
+                                      );
+                                      setInnerGrid(grid, solution);
+                                      setTempSolvePath([]);
+                                    }}
+                                  >
+                                    Reset progress
+                                  </button>
+                                  <button
+                                    type="button"
+                                    className="btn btn-primary"
+                                    onClick={() => {
+                                      setInitialState(null);
+                                    }}
+                                  >
+                                    Save and exit
+                                  </button>
+                                </>
+                              );
+                            }}
+                          </GridConsumer>
+                        </PerfectionScreen>
+                      </GridContext>
+                    </GridStateContext>
+                  </EditContext>
+                </DisplayContext>
+              </EmbedContext>
+            )}
+          </div>
+        </dialog>
+      );
+    }
+  )
+);
 
 export const type = undefined;